refactor: implement MVVM architecture for ExploreView
